Revert getSublabelAndPrice
